Fix support for esm
